backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
firebase-admin
google-cloud-storage
google-cloud-bigquery
scipy
numpy
pandas
httpx
reportlab
python-dotenv
pydantic>=2.0
python-multipart
